requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-docx==0.8.11
pdfplumber==0.11.7
pandas==2.2.2
numpy==1.26.4
spacy==3.8.1
dateparser==1.1.8
scikit-learn==1.3.2
python-magic==0.4.27

openpyxl==3.1.3
python-dotenv==1.0.0
